# sql: optimize the prepared statements cache