--- a/pkg/server/serverpb/BUILD.bazel
+++ b/pkg/server/serverpb/BUILD.bazel
@@ -72,6 +72,7 @@ go_proto_library(
         # NB: The grpc-gateway compiler injects a dependency on the descriptor
         # package that Gazelle isn't prepared to deal with.
         "@com_github_golang_protobuf//descriptor:go_default_library_gen",  # keep
+        "@org_golang_google_genproto_googleapis_api//annotations:go_default_library",
         "//pkg/build",
         "//pkg/clusterversion",
         "//pkg/config/zonepb",
@@ -94,7 +95,6 @@ go_proto_library(
         "//pkg/util/log/logpb",
         "//pkg/util/metric",
         "//pkg/util/tracing/tracingpb",
-        "@org_golang_google_genproto_googleapis_api//annotations:go_default_library",
     ],
 )