--- a/pkg/sql/conn_executor.go
+++ b/pkg/sql/conn_executor.go
@@ -9,7 +9,6 @@ import (
 	"context"
 	"fmt"
 	"io"
-	"maps"
 	"math"
 	"math/rand"
 	"strings"
@@ -79,6 +78,7 @@ import (
 	"github.com/cockroachdb/cockroach/pkg/util/errorutil"
 	"github.com/cockroachdb/cockroach/pkg/util/fsm"
 	"github.com/cockroachdb/cockroach/pkg/util/hlc"
+	"github.com/cockroachdb/cockroach/pkg/util/humanizeutil"
 	"github.com/cockroachdb/cockroach/pkg/util/log"
 	"github.com/cockroachdb/cockroach/pkg/util/log/logcrash"
 	"github.com/cockroachdb/cockroach/pkg/util/log/severity"
@@ -1255,16 +1255,9 @@ func (s *Server) newConnExecutor(
 	}
 
 	ex.extraTxnState.underOuterTxn = underOuterTxn
-	ex.extraTxnState.prepStmtsNamespace = prepStmtNamespace{
-		prepStmts:    make(map[string]*prep.Statement),
-		prepStmtsLRU: make(map[string]struct{ prev, next string }),
-		portals:      make(map[string]PreparedPortal),
-	}
-	ex.extraTxnState.prepStmtsNamespaceAtTxnRewindPos = prepStmtNamespace{
-		prepStmts:    make(map[string]*prep.Statement),
-		prepStmtsLRU: make(map[string]struct{ prev, next string }),
-		portals:      make(map[string]PreparedPortal),
-	}
+	ex.extraTxnState.prepStmtsNamespace.prepStmts.Init()
+	ex.extraTxnState.prepStmtsNamespace.portals = make(map[string]PreparedPortal)
+	ex.extraTxnState.prepStmtsNamespace.portalsSnapshot = make(map[string]PreparedPortal)
 	ex.extraTxnState.prepStmtsNamespaceMemAcc = ex.sessionMon.MakeBoundAccount()
 	dsdp := catsessiondata.NewDescriptorSessionDataStackProvider(sdMutIterator.sds)
 	ex.extraTxnState.descCollection = s.cfg.CollectionFactory.NewCollection(
@@ -1358,9 +1351,6 @@ func (ex *connExecutor) close(ctx context.Context, closeType closeType) {
 	ex.extraTxnState.prepStmtsNamespace.closeAllPortals(
 		ctx, &ex.extraTxnState.prepStmtsNamespaceMemAcc,
 	)
-	ex.extraTxnState.prepStmtsNamespaceAtTxnRewindPos.closeAllPortals(
-		ctx, &ex.extraTxnState.prepStmtsNamespaceMemAcc,
-	)
 	if err := ex.extraTxnState.sqlCursors.closeAll(&ex.planner, cursorCloseForExplicitClose); err != nil {
 		log.Warningf(ctx, "error closing cursors: %v", err)
 	}
@@ -1419,10 +1409,7 @@ func (ex *connExecutor) close(ctx context.Context, closeType closeType) {
 	if closeType != panicClose {
 		// Close all statements and prepared portals. The cursors have already been
 		// closed.
-		ex.extraTxnState.prepStmtsNamespace.resetToEmpty(
-			ctx, &ex.extraTxnState.prepStmtsNamespaceMemAcc,
-		)
-		ex.extraTxnState.prepStmtsNamespaceAtTxnRewindPos.resetToEmpty(
+		ex.extraTxnState.prepStmtsNamespace.clear(
 			ctx, &ex.extraTxnState.prepStmtsNamespaceMemAcc,
 		)
 		ex.extraTxnState.prepStmtsNamespaceMemAcc.Close(ctx)
@@ -1587,26 +1574,18 @@ type connExecutor struct {
 		// Set via setTxnRewindPos().
 		txnRewindPos CmdPos
 
-		// prepStmtNamespace contains the prepared statements and portals that the
-		// session currently has access to.
-		// Portals are bound to a transaction and they're all destroyed once the
-		// transaction finishes.
-		// Prepared statements are not transactional and so it's a bit weird that
-		// they're part of extraTxnState, but it's convenient to put them here
-		// because they need the same kind of "snapshoting" as the portals (see
-		// prepStmtsNamespaceAtTxnRewindPos).
+		// prepStmtNamespace contains the portals and prepared statements that
+		// the session currently has access to. To interact properly with
+		// automatic transaction retries, we need the ability to restore portals
+		// and prepared statements as they were at the previous txnRewindPos. To
+		// achieve this for portals we take a snapshot even time txnRewindPos is
+		// advanced. For prepared statements we use the Commit and Rewind
+		// methods of prep.Cache.
+		//
+		// Prepared statements are not transactional and so it's a bit weird
+		// that they're part of extraTxnState, but it's convenient to put them
+		// here because they need the same kind of snapshotting as the portals.
 		prepStmtsNamespace prepStmtNamespace
-
-		// prepStmtsNamespaceAtTxnRewindPos is a snapshot of the prep stmts/portals
-		// (ex.prepStmtsNamespace) before processing the command at position
-		// txnRewindPos.
-		// Here's the deal: prepared statements are not transactional, but they do
-		// need to interact properly with automatic retries (i.e. rewinding the
-		// command buffer). When doing a rewind, we need to be able to restore the
-		// prep stmts as they were. We do this by taking a snapshot every time
-		// txnRewindPos is advanced. Prepared statements are shared between the two
-		// collections, but these collections are periodically reconciled.
-		prepStmtsNamespaceAtTxnRewindPos prepStmtNamespace
 
 		// prepStmtsNamespaceMemAcc is the memory account that is shared
 		// between prepStmtsNamespace and prepStmtsNamespaceAtTxnRewindPos. It
@@ -1940,23 +1919,14 @@ func (ch *ctxHolder) unhijack() {
 type prepStmtNamespace struct {
 	// prepStmts contains the prepared statements currently available on the
 	// session.
-	prepStmts map[string]*prep.Statement
-	// prepStmtsLRU is a circular doubly-linked list containing the prepared
-	// statement names ordered by most recent access (needed to determine
-	// evictions when prepared_statements_cache_size is set). There is a special
-	// entry for the empty string which is both the head and tail of the
-	// list. (Consequently, if it exists, the actual prepared statement for the
-	// empty string does not have an entry in this list and cannot be evicted.)
-	prepStmtsLRU map[string]struct{ prev, next string }
-	// prepStmtsLRUAlloc is the total amount of memory allocated for prepared
-	// statements in prepStmtsLRU. This will sometimes be less than
-	// ex.sessionPreparedMon.AllocBytes() because refcounting causes us to hold
-	// onto more prep.Statements than are currently in the LRU list.
-	prepStmtsLRUAlloc int64
+	prepStmts prep.Cache
 	// portals contains the portals currently available on the session. Note
 	// that PreparedPortal.accountForCopy needs to be called if a copy of a
 	// PreparedPortal is retained.
 	portals map[string]PreparedPortal
+	// portalsSnapshot is a snapshot of portals at the time of the last call to
+	// commit.
+	portalsSnapshot map[string]PreparedPortal
 }
 
 // HasActivePortals returns true if there are portals in the session.
@@ -1970,95 +1940,34 @@ func (ns *prepStmtNamespace) HasPortal(s string) bool {
 	return ok
 }
 
-const prepStmtsLRUHead = ""
-const prepStmtsLRUTail = ""
-
-// addLRUEntry adds a new prepared statement name to the LRU list. It is an
-// error to re-add an existing name to the LRU list.
-func (ns *prepStmtNamespace) addLRUEntry(name string, alloc int64) {
-	if name == prepStmtsLRUHead {
-		return
-	}
-	if _, ok := ns.prepStmtsLRU[name]; ok {
-		// Assert that we're not re-adding an existing name to the LRU list.
-		panic(errors.AssertionFailedf(
-			"prepStmtsLRU unexpected existing entry (%s): %v", name, ns.prepStmtsLRU,
-		))
-	}
-	var this struct{ prev, next string }
-	this.prev = prepStmtsLRUHead
-	// Note: must do this serially in case head and next are the same entry.
-	head := ns.prepStmtsLRU[this.prev]
-	this.next = head.next
-	head.next = name
-	ns.prepStmtsLRU[prepStmtsLRUHead] = head
-	next, ok := ns.prepStmtsLRU[this.next]
-	if !ok || next.prev != prepStmtsLRUHead {
-		// Assert that the chain isn't broken before we modify it.
-		panic(errors.AssertionFailedf(
-			"prepStmtsLRU head entry not correct (%s): %v", this.next, ns.prepStmtsLRU,
-		))
-	}
-	next.prev = name
-	ns.prepStmtsLRU[this.next] = next
-	ns.prepStmtsLRU[name] = this
-	ns.prepStmtsLRUAlloc += alloc
-}
-
-// delLRUEntry removes a prepared statement name from the LRU list. (It is not an
-// error to remove a non-existent prepared statement.)
-func (ns *prepStmtNamespace) delLRUEntry(name string, alloc int64) {
-	if name == prepStmtsLRUHead {
-		return
-	}
-	this, ok := ns.prepStmtsLRU[name]
-	if !ok {
-		// Not an error to remove a non-existent prepared statement.
-		return
-	}
-	// Note: must do this serially in case prev and next are the same entry.
-	prev, ok := ns.prepStmtsLRU[this.prev]
-	if !ok || prev.next != name {
-		// Assert that the chain isn't broken before we modify it.
-		panic(errors.AssertionFailedf(
-			"prepStmtsLRU prev entry not correct (%s): %v", this.prev, ns.prepStmtsLRU,
-		))
-	}
-	prev.next = this.next
-	ns.prepStmtsLRU[this.prev] = prev
-	next, ok := ns.prepStmtsLRU[this.next]
-	if !ok || next.prev != name {
-		// Assert that the chain isn't broken before we modify it.
-		panic(errors.AssertionFailedf(
-			"prepStmtsLRU next entry not correct (%s): %v", this.next, ns.prepStmtsLRU,
-		))
-	}
-	next.prev = this.prev
-	ns.prepStmtsLRU[this.next] = next
-	delete(ns.prepStmtsLRU, name)
-	ns.prepStmtsLRUAlloc -= alloc
-}
-
-// touchLRUEntry moves an existing prepared statement to the front of the LRU
-// list.
-func (ns *prepStmtNamespace) touchLRUEntry(name string) {
-	if name == prepStmtsLRUHead {
-		return
-	}
-	if ns.prepStmtsLRU[prepStmtsLRUHead].next == name {
-		// Already at the front of the list.
-		return
-	}
-	ns.delLRUEntry(name, 0)
-	ns.addLRUEntry(name, 0)
-}
-
 func (ns *prepStmtNamespace) closeAllPortals(
 	ctx context.Context, prepStmtsNamespaceMemAcc *mon.BoundAccount,
 ) {
 	for name, p := range ns.portals {
 		p.close(ctx, prepStmtsNamespaceMemAcc, name)
 		delete(ns.portals, name)
+	}
+	for name, p := range ns.portalsSnapshot {
+		p.close(ctx, prepStmtsNamespaceMemAcc, name)
+		delete(ns.portalsSnapshot, name)
+	}
+}
+
+func (ns *prepStmtNamespace) closePortals(
+	ctx context.Context, prepStmtsNamespaceMemAcc *mon.BoundAccount,
+) {
+	for name, p := range ns.portals {
+		p.close(ctx, prepStmtsNamespaceMemAcc, name)
+		delete(ns.portals, name)
+	}
+}
+
+func (ns *prepStmtNamespace) closeSnapshotPortals(
+	ctx context.Context, prepStmtsNamespaceMemAcc *mon.BoundAccount,
+) {
+	for name, p := range ns.portalsSnapshot {
+		p.close(ctx, prepStmtsNamespaceMemAcc, name)
+		delete(ns.portalsSnapshot, name)
 	}
 }
 
@@ -2074,14 +1983,18 @@ func (ns *prepStmtNamespace) closeAllPausablePortals(
 }
 
 // MigratablePreparedStatements returns a mapping of all prepared statements.
-func (ns *prepStmtNamespace) MigratablePreparedStatements() []sessiondatapb.MigratableSession_PreparedStatement {
-	ret := make([]sessiondatapb.MigratableSession_PreparedStatement, 0, len(ns.prepStmts))
+func (ns *prepStmtNamespace) MigratablePreparedStatements() (
+	[]sessiondatapb.MigratableSession_PreparedStatement,
+	error,
+) {
+	if ns.prepStmts.Dirty() {
+		return nil, errors.AssertionFailedf("cannot serialize dirty prepared statements cache")
+	}
 
 	// Serialize prepared statements from least-recently used to most-recently
 	// used, so that we build the LRU list correctly when deserializing.
-	for e, ok := ns.prepStmtsLRU[prepStmtsLRUTail]; ok && e.prev != prepStmtsLRUHead; e, ok = ns.prepStmtsLRU[e.prev] {
-		name := e.prev
-		stmt := ns.prepStmts[name]
+	ret := make([]sessiondatapb.MigratableSession_PreparedStatement, 0, ns.prepStmts.Len())
+	ns.prepStmts.ForEachLRU(func(name string, stmt *prep.Statement) {
 		ret = append(
 			ret,
 			sessiondatapb.MigratableSession_PreparedStatement{
@@ -2090,33 +2003,22 @@ func (ns *prepStmtNamespace) MigratablePreparedStatements() []sessiondatapb.Migr
 				SQL:                  stmt.SQL,
 			},
 		)
-	}
-	// Finally, serialize the anonymous prepared statement (if it exists).
-	if stmt, ok := ns.prepStmts[""]; ok {
-		ret = append(
-			ret,
-			sessiondatapb.MigratableSession_PreparedStatement{
-				Name:                 "",
-				PlaceholderTypeHints: stmt.InferredTypes,
-				SQL:                  stmt.SQL,
-			},
-		)
-	}
-
-	return ret
+	})
+	return ret, nil
 }
 
 func (ns *prepStmtNamespace) String() string {
 	var sb strings.Builder
 	sb.WriteString("Prep stmts: ")
-	// Put the anonymous prepared statement first (if it exists).
-	if _, ok := ns.prepStmts[""]; ok {
-		sb.WriteString("\"\" ")
+	if ns.prepStmts.Dirty() {
+		sb.WriteString("<dirty>")
+	} else {
+		ns.prepStmts.ForEachLRU(func(name string, stmt *prep.Statement) {
+			sb.WriteString(name)
+			sb.WriteByte(' ')
+		})
 	}
-	for e, ok := ns.prepStmtsLRU[prepStmtsLRUHead]; ok && e.next != prepStmtsLRUTail; e, ok = ns.prepStmtsLRU[e.next] {
-		sb.WriteString(e.next + " ")
-	}
-	fmt.Fprintf(&sb, "LRU alloc: %d ", ns.prepStmtsLRUAlloc)
+	fmt.Fprintf(&sb, "Size: %d ", ns.prepStmts.Size())
 	sb.WriteString("Portals: ")
 	for name := range ns.portals {
 		sb.WriteString(name + " ")
@@ -2124,48 +2026,48 @@ func (ns *prepStmtNamespace) String() string {
 	return sb.String()
 }
 
-// resetToEmpty deallocates the prepStmtNamespace.
-func (ns *prepStmtNamespace) resetToEmpty(
+// clear empties the prepStmtNamespace.
+func (ns *prepStmtNamespace) clear(
 	ctx context.Context, prepStmtsNamespaceMemAcc *mon.BoundAccount,
 ) {
-	// No errors could occur since we're releasing the resources.
-	_ = ns.resetTo(ctx, &prepStmtNamespace{}, prepStmtsNamespaceMemAcc)
+	// Commit the prepared statements cache to ensure that any statements that
+	// have been removed but not committed are not leaked.
+	// TODO(mgartner): Make a Clear method of prep.Cache to make this automatic.
+	ns.prepStmts.Commit(ctx, 0)
+	ns.prepStmts.ForEach(func(name string, stmt *prep.Statement) {
+		stmt.DecRef(ctx)
+	})
+	ns.prepStmts.Init()
+	ns.closeAllPortals(ctx, prepStmtsNamespaceMemAcc)
 }
 
-// resetTo resets a namespace to equate another one (`to`). All the receiver's
-// references are released and all the to's references are duplicated.
-//
-// An empty `to` can be passed in to deallocate everything.
-//
-// It can only return an error if we've reached the memory limit and had to make
-// a copy of portals.
-func (ns *prepStmtNamespace) resetTo(
-	ctx context.Context, to *prepStmtNamespace, prepStmtsNamespaceMemAcc *mon.BoundAccount,
+func (ns *prepStmtNamespace) commit(
+	ctx context.Context, maxSize int64, prepStmtsNamespaceMemAcc *mon.BoundAccount,
 ) error {
-	// Reset prepStmts.
-	if !maps.Equal(ns.prepStmts, to.prepStmts) {
-		for name, ps := range ns.prepStmts {
-			ps.DecRef(ctx)
-			delete(ns.prepStmts, name)
-		}
-		for name, ps := range to.prepStmts {
-			ps.IncRef(ctx)
-			ns.prepStmts[name] = ps
-		}
+	evicted := ns.prepStmts.Commit(ctx, maxSize)
+	for i := range evicted {
+		log.VEventf(
+			ctx, 1,
+			"prepared statements are using more than prepared_statements_cache_size (%s), "+
+				"automatically deallocating %s", string(humanizeutil.IBytes(maxSize)), evicted[i],
+		)
 	}
-
-	// Reset prepStmtsLRU.
-	if !maps.Equal(ns.prepStmtsLRU, to.prepStmtsLRU) {
-		clear(ns.prepStmtsLRU)
-		maps.Copy(ns.prepStmtsLRU, to.prepStmtsLRU)
+	ns.closeSnapshotPortals(ctx, prepStmtsNamespaceMemAcc)
+	for name, p := range ns.portals {
+		if err := p.accountForCopy(ctx, prepStmtsNamespaceMemAcc, name); err != nil {
+			return err
+		}
+		ns.portalsSnapshot[name] = p
 	}
+	return nil
+}
 
-	// Reset prepStmtsLRUAlloc.
-	ns.prepStmtsLRUAlloc = to.prepStmtsLRUAlloc
-
-	// Reset portals.
-	ns.closeAllPortals(ctx, prepStmtsNamespaceMemAcc)
-	for name, p := range to.portals {
+func (ns *prepStmtNamespace) rewind(
+	ctx context.Context, prepStmtsNamespaceMemAcc *mon.BoundAccount,
+) error {
+	ns.prepStmts.Rewind(ctx)
+	ns.closePortals(ctx, prepStmtsNamespaceMemAcc)
+	for name, p := range ns.portalsSnapshot {
 		if err := p.accountForCopy(ctx, prepStmtsNamespaceMemAcc, name); err != nil {
 			return err
 		}
@@ -2202,7 +2104,7 @@ func (ex *connExecutor) resetExtraTxnState(ctx context.Context, ev txnEvent, pay
 	}
 
 	// Close all portals.
-	ex.extraTxnState.prepStmtsNamespace.closeAllPortals(
+	ex.extraTxnState.prepStmtsNamespace.closePortals(
 		ctx, &ex.extraTxnState.prepStmtsNamespaceMemAcc,
 	)
 
@@ -2222,7 +2124,7 @@ func (ex *connExecutor) resetExtraTxnState(ctx context.Context, ev txnEvent, pay
 
 	switch ev.eventType {
 	case txnCommit, txnRollback, txnPrepare:
-		ex.extraTxnState.prepStmtsNamespaceAtTxnRewindPos.closeAllPortals(
+		ex.extraTxnState.prepStmtsNamespace.closeSnapshotPortals(
 			ctx, &ex.extraTxnState.prepStmtsNamespaceMemAcc,
 		)
 		ex.extraTxnState.savepoints.clear()
@@ -3472,16 +3374,17 @@ func stmtHasNoData(stmt tree.Statement) bool {
 // commitPrepStmtNamespace deallocates everything in
 // prepStmtsNamespaceAtTxnRewindPos that's not part of prepStmtsNamespace.
 func (ex *connExecutor) commitPrepStmtNamespace(ctx context.Context) error {
-	return ex.extraTxnState.prepStmtsNamespaceAtTxnRewindPos.resetTo(
-		ctx, &ex.extraTxnState.prepStmtsNamespace, &ex.extraTxnState.prepStmtsNamespaceMemAcc,
+	cacheSize := ex.sessionData().PreparedStatementsCacheSize
+	return ex.extraTxnState.prepStmtsNamespace.commit(
+		ctx, cacheSize, &ex.extraTxnState.prepStmtsNamespaceMemAcc,
 	)
 }
 
-// commitPrepStmtNamespace deallocates everything in prepStmtsNamespace that's
+// rewindPrepStmtNamespace deallocates everything in prepStmtsNamespace that's
 // not part of prepStmtsNamespaceAtTxnRewindPos.
 func (ex *connExecutor) rewindPrepStmtNamespace(ctx context.Context) error {
-	return ex.extraTxnState.prepStmtsNamespace.resetTo(
-		ctx, &ex.extraTxnState.prepStmtsNamespaceAtTxnRewindPos, &ex.extraTxnState.prepStmtsNamespaceMemAcc,
+	return ex.extraTxnState.prepStmtsNamespace.rewind(
+		ctx, &ex.extraTxnState.prepStmtsNamespaceMemAcc,
 	)
 }
 
@@ -4206,7 +4109,7 @@ func (ex *connExecutor) txnStateTransitionsApplyWrapper(
 		ex.resetExtraTxnState(ex.Ctx(), advInfo.txnEvent, payloadErr)
 		// Since we're finalizing the SQL transaction (commit, rollback, prepare),
 		// there's no need to keep the prepared stmts for a txn rewind.
-		ex.extraTxnState.prepStmtsNamespaceAtTxnRewindPos.closeAllPortals(
+		ex.extraTxnState.prepStmtsNamespace.closeSnapshotPortals(
 			ex.Ctx(), &ex.extraTxnState.prepStmtsNamespaceMemAcc,
 		)
 		ex.resetPlanner(ex.Ctx(), &ex.planner, nil, ex.server.cfg.Clock.PhysicalTime())
@@ -4841,37 +4744,33 @@ var _ preparedStatementsAccessor = connExPrepStmtsAccessor{}
 
 // List is part of the preparedStatementsAccessor interface.
 func (ps connExPrepStmtsAccessor) List() map[string]*prep.Statement {
-	// Return a copy of the data, to prevent modification of the map.
-	stmts := ps.ex.extraTxnState.prepStmtsNamespace.prepStmts
-	ret := make(map[string]*prep.Statement, len(stmts))
-	for key, stmt := range stmts {
-		ret[key] = stmt
-	}
+	prepStmts := ps.ex.extraTxnState.prepStmtsNamespace.prepStmts
+	ret := make(map[string]*prep.Statement, prepStmts.Len())
+	prepStmts.ForEach(func(name string, stmt *prep.Statement) {
+		ret[name] = stmt
+	})
 	return ret
 }
 
 // Get is part of the preparedStatementsAccessor interface.
-func (ps connExPrepStmtsAccessor) Get(name string, touchLRU bool) (*prep.Statement, bool) {
-	s, ok := ps.ex.extraTxnState.prepStmtsNamespace.prepStmts[name]
-	if ok && touchLRU {
-		ps.ex.extraTxnState.prepStmtsNamespace.touchLRUEntry(name)
-	}
-	return s, ok
+func (ps connExPrepStmtsAccessor) Get(name string) (*prep.Statement, bool) {
+	prepStmts := &ps.ex.extraTxnState.prepStmtsNamespace.prepStmts
+	return prepStmts.Get(name)
 }
 
 // Delete is part of the preparedStatementsAccessor interface.
 func (ps connExPrepStmtsAccessor) Delete(ctx context.Context, name string) bool {
-	_, ok := ps.Get(name, false /* touchLRU */)
-	if !ok {
+	prepStmts := &ps.ex.extraTxnState.prepStmtsNamespace.prepStmts
+	if !prepStmts.Has(name) {
 		return false
 	}
-	ps.ex.deletePreparedStmt(ctx, name)
+	prepStmts.Remove(name)
 	return true
 }
 
 // DeleteAll is part of the preparedStatementsAccessor interface.
 func (ps connExPrepStmtsAccessor) DeleteAll(ctx context.Context) {
-	ps.ex.extraTxnState.prepStmtsNamespace.resetToEmpty(
+	ps.ex.extraTxnState.prepStmtsNamespace.clear(
 		ctx, &ps.ex.extraTxnState.prepStmtsNamespaceMemAcc,
 	)
 }