--- a/pkg/sql/prep/statement_test.go
+++ b/pkg/sql/prep/statement_test.go
@@ -3,7 +3,7 @@
 // Use of this software is governed by the CockroachDB Software License
 // included in the /LICENSE file.
 
-package sql
+package prep
 
 import (
 	"testing"