… pkg/sql/conn_executor_exec.go
… pkg/sql/faketreeeval/evalctx.go
… pkg/sql/logictest/testdata/logic_test/prepare
… pkg/sql/pg_catalog.go
… pkg/sql/plan_opt.go
… pkg/sql/plan_opt_test.go
… pkg/sql/planner.go
… pkg/sql/prep/BUILD.bazel
… pkg/sql/prep/cache.go
… pkg/sql/prep/cache_test.go
… pkg/sql/prep/metadata.go
… pkg/sql/prep/statement.go
… pkg/sql/prepared_stmt.go
… pkg/sql/querycache/BUILD.bazel
… pkg/sql/querycache/query_cache.go
… pkg/sql/querycache/query_cache_test.go
… pkg/sql/sem/eval/deps.go
… pkg/sql/session_state.go
… pkg/sql/statement.go
… pkg/sql/two_phase_commit.go