#### querycache,prep: move PrepareMetadata to new prep package

This is a mechanical change that moves `querycache.PrepareMetadata` to a
new `prep` package in preparation for future changes.

Release note: None

#### sql,prep: move PreparedStatement to prep package

This is a mechanical change the moves `sql.PreparedStatement` to the
`prep` package in preparation for future changes.

Release note: None

#### prep: add Cache

A new `prep.Cache` data structure has been added which maps prepared
statement names to `*prep.Statement`s. `Cache` provides a concise API
and it is optimized for the most common interactions in most workloads:

  1. Fetching prepared statements with `Get`.
  2. Snapshotting the `Cache` with `Commit` when no changes have been
     applied since the last `Commit`.

In order to facilitate automatic transaction retries which rewind the
command buffer, changes to the prepared statements cache must be undone.
We currently achieve this using two maps: one up-to-date map and the
other a snapshot of the prepared statements after the last transaction
committed. Before retrying, we clear the up-to-date copy of the map and
copy the contents of the snapshot into it, effectively undoing and
changes to the cache in the current transaction. If the transaction is
successful, we clear the snapshot and copy the up-to-date map into it,
effectively creating a new snapshot.

This approach is expensive even if the prepared statements cache is not
altered. If the cache hasn't changed, we can avoid clearing and copying
into the snapshot, but we still need to check for map equality to
determine if there have been changes. This equality check has time
complexity proportional to the number of cached prepared statements.

Also, there are two other maps we use implement an LRU eviction policy.
They keep track of the ordering of most recent access for each statement
(one is up-to-date and the other is a snapshot). Every time a prepared
statement is executed that was not the previously executed statement,
the most recent access ordering changes. Therefore, we can rarely avoid
the clear and copy of the snapshot like we can with the other maps.

The current implementation also lacks a clear API - the four maps are
separate objects and directly accessed across the `sql` pacakge and
prone to misuse.

`Cache` achieves similar behavior more efficiently with a clear API. It
provides `Add`, `Get`, `Remove`, and other methods with obvious purpose.
The `Rewind` and `Commit` methods facilitate automatic retries. `Commit`
is a constant-time operation if there have been no changes to the
`Cache` since the last `Commit`. See the inline comments for more
details.

The behavior of `Cache`, when used, will vary slightly from the current
behavior. Prior to this commit prepared statements were evicted as soon
new statements were added. Now, prepared statements are evicted when the
cache is committed, i.e., when statements succeed (or fail without a
retriable error) and we know that the command buffer will not be
rewound. This difference in behavior should have no meaningful impact
for users.

Fixes #143732

Release note (performance improvement): Prepared statements are now more
efficiently cached.

#### sql: replace prepared statement maps with prep.Cache

Four maps using to implement the prepared statements cache have been
repalced with `prep.Cache`. Additional, a snapshot of a connection's
portals is now kept in a map in `prepStmtNamespace` so that
`prepStmtNamespace` can implement `commit` and `rewind` methods.

Release note (performance improvement): Prepared statements are now more
efficiently cached.

#### prep: decrement statement reference counts in Cache.Init

Prepared statement reference counts are now decremented in `Cache.Init`
if the cache is non-empty. This makes it easier and safe to clear a
`Cache` with `Init`.

Also, some minor changes have been made to simplify the test oracle.

Release note: None