@mgartner 2025-04-08
Some micro-benchmark results:

```
name                                  old time/op    new time/op    delta
Sysbench/SQL/3node/oltp_point_select     176µs ± 6%     156µs ± 5%  -11.75%  (p=0.008 n=5+5)

name                                  old errs/op    new errs/op    delta
Sysbench/SQL/3node/oltp_point_select      0.00           0.00          ~     (all equal)

name                                  old alloc/op   new alloc/op   delta
Sysbench/SQL/3node/oltp_point_select    25.6kB ± 1%    25.2kB ± 1%   -1.45%  (p=0.008 n=5+5)

name                                  old allocs/op  new allocs/op  delta
Sysbench/SQL/3node/oltp_point_select       218 ± 1%       217 ± 0%     ~     (p=0.095 n=5+4)
```

I'll collect more benchmark results soon.