@mgartner 2025-04-10
Unfortunately, I could not measure any improvement in throughput in a full oltp_read_write test.